+++ common/crc32.h
@@ -15,7 +15,7 @@ extern "C" {
 /// @param buf The data to be hashed
 /// @param len The size of the data
 /// @return CRC-32 hash
-u32 crc32buf(const u8* buf, u32 len);
+u32 crc32buf(const u8* buf, u64 len);
 
 /// @brief Compute the CRC-32C hash of a buffer
 ///
@@ -27,14 +27,14 @@ u32 crc32buf(const u8* buf, u32 len);
 /// @param buf The data to be hashed
 /// @param len The size of the data
 /// @return CRC-32C hash
-u32 crc32c(const u8* buf, u32 len);
+u32 crc32c(const u8* buf, u64 len);
 
 // SSE-only implementation of CRC32C, exposed only for unit testing
-u32 sse_crc32c(const u8* buf, u32 len);
+u32 sse_crc32c(const u8* buf, u64 len);
 
 // Software-only implementation of CRC32C. Please don't call directly, the
 // generic version will do SSE for you if possible.
-u32 software_crc32c(const u8* buf, u32 len);
+u32 software_crc32c(const u8* buf, u64 len);
 
 // Always calls the fastest available variant of CRC32.
 #define crc32fast crc32c
